+++ src/Sulu/Component/Content/Document/Subscriber/StructureSubscriber.php
@@ -121,6 +121,7 @@ class StructureSubscriber implements EventSubscriberInterface
                 'load_ghost_content' => true,
                 'clear_missing_content' => false,
                 'ignore_required' => false,
+                'structure_type' => null,
             ]
         );
         $options->setAllowedTypes('load_ghost_content', 'bool');
@@ -177,15 +178,8 @@ class StructureSubscriber implements EventSubscriberInterface
             return;
         }
 
-        $node = $event->getNode();
-        $propertyName = $this->getStructureTypePropertyName($document, $event->getLocale());
-        $structureType = $node->getPropertyValueWithDefault($propertyName, null);
-
         $rehydrate = $event->getOption('rehydrate');
-        if (!$structureType && $rehydrate) {
-            $structureType = $this->getDefaultStructureType($document);
-        }
-
+        $structureType = $this->getStructureType($event, $document, $rehydrate);
         $document->setStructureType($structureType);
 
         if (false === $event->getOption('load_ghost_content', false)) {
@@ -233,6 +227,31 @@ class StructureSubscriber implements EventSubscriberInterface
             $this->getStructureTypePropertyName($document, $locale),
             $document->getStructureType()
         );
+    }
+
+    /**
+     * @param AbstractMappingEvent $event
+     * @param StructureBehavior $document
+     * @param bool $rehydrate
+     *
+     * @return string
+     */
+    private function getStructureType(AbstractMappingEvent $event, StructureBehavior $document, $rehydrate)
+    {
+        $structureType = $event->getOption('structure_type');
+        if ($structureType) {
+            return $structureType;
+        }
+
+        $node = $event->getNode();
+        $propertyName = $this->getStructureTypePropertyName($document, $event->getLocale());
+        $structureType = $node->getPropertyValueWithDefault($propertyName, null);
+
+        if (!$structureType && $rehydrate) {
+            return $this->getDefaultStructureType($document);
+        }
+
+        return $structureType;
     }
 
     /**