+++ src/Sulu/Component/Content/Tests/Unit/Document/Subscriber/StructureSubscriberTest.php
@@ -282,6 +282,7 @@ class StructureSubscriberTest extends SubscriberTestCase
         $this->hydrateEvent->getLocale()->willReturn('fr');
         $this->hydrateEvent->getOption('load_ghost_content', false)->willReturn(true);
         $this->hydrateEvent->getOption('rehydrate')->willReturn(false);
+        $this->hydrateEvent->getOption('structure_type')->willReturn(null);
 
         // set the structure type
         $this->encoder->contentName('template')->willReturn('i18n:fr-template');
@@ -304,6 +305,7 @@ class StructureSubscriberTest extends SubscriberTestCase
         $this->hydrateEvent->getLocale()->willReturn('fr');
         $this->hydrateEvent->getOption('load_ghost_content', false)->willReturn(true);
         $this->hydrateEvent->getOption('rehydrate')->willReturn(false);
+        $this->hydrateEvent->getOption('structure_type')->willReturn(null);
 
         // set the structure type
         $this->encoder->contentName('template')->willReturn('i18n:fr-template');
@@ -315,6 +317,29 @@ class StructureSubscriberTest extends SubscriberTestCase
         // set the property container
         $this->subscriber->handleHydrate($this->hydrateEvent->reveal());
         $this->accessor->set('structure', Argument::type(Structure::class))->shouldHaveBeenCalled();
+    }
+
+    /**
+     * It should load data the correct structure when template is given.
+     */
+    public function testHydrateOtherTemplate()
+    {
+        $this->hydrateEvent->getDocument()->willReturn($this->document->reveal());
+        $this->hydrateEvent->getNode()->willReturn($this->node->reveal());
+        $this->hydrateEvent->getLocale()->willReturn('fr');
+        $this->hydrateEvent->getOption('load_ghost_content', false)->willReturn(true);
+        $this->hydrateEvent->getOption('rehydrate')->willReturn(false);
+        $this->hydrateEvent->getOption('structure_type')->willReturn('test');
+
+        // set the structure type
+        $this->encoder->contentName('template')->willReturn('i18n:fr-template');
+        $this->node->getPropertyValueWithDefault('i18n:fr-template', null)->willReturn('foobar');
+
+        $this->document->setStructureType('test')->shouldBeCalled();
+
+        // set the property container
+        $this->subscriber->handleHydrate($this->hydrateEvent->reveal());
+        $this->accessor->set('structure', Argument::type(ManagedStructure::class))->shouldHaveBeenCalled();
     }
 
     /**
@@ -333,6 +358,7 @@ class StructureSubscriberTest extends SubscriberTestCase
         $this->hydrateEvent->getLocale()->willReturn('fr');
         $this->hydrateEvent->getOption('load_ghost_content', false)->willReturn(true);
         $this->hydrateEvent->getOption('rehydrate')->willReturn(true);
+        $this->hydrateEvent->getOption('structure_type')->willReturn(null);
 
         // set the structure type
         $this->encoder->contentName('template')->willReturn('i18n:fr-template');
@@ -357,6 +383,7 @@ class StructureSubscriberTest extends SubscriberTestCase
         $this->hydrateEvent->getLocale()->willReturn('fr');
         $this->hydrateEvent->getOption('load_ghost_content', false)->willReturn(true);
         $this->hydrateEvent->getOption('rehydrate')->willReturn(false);
+        $this->hydrateEvent->getOption('structure_type')->willReturn(null);
 
         // set the structure type
         $this->encoder->contentName('template')->willReturn('i18n:fr-template');
@@ -384,6 +411,8 @@ class StructureSubscriberTest extends SubscriberTestCase
         $this->hydrateEvent->getLocale()->willReturn('fr');
         $this->hydrateEvent->getOption('load_ghost_content', false)->willReturn(true);
         $this->hydrateEvent->getOption('rehydrate')->willReturn(true);
+        $this->hydrateEvent->getOption('structure_type')->willReturn(null);
+
         $this->inspector->getWebspace($this->document->reveal())->willReturn('sulu_io');
         $this->inspector->getMetadata($this->document->reveal())->willReturn($metadata->reveal());